ROX-23287: compliance refactor result service (#10925)
Refactor of results service.
